Make panesPerWindow consistent with commands passed to smux.
Previously, we would pass only the first set of commands to smux in the
case where noCreate was passed, but we would pass the full length of
commands. This inconsistency resulted in smux creating a bunch of empty
panes. This commit makes it so that commands passed are consistent with
length passed.

Bumped version to 0.1.5.