Add option --all_namespaces.
This commit changes the default behavior of kmux to use the "default"
namespace instead of all namespaces. It also introduces the
all_namespaces flag for specifying that the user wants to examine pods
and deployments and replica sets in all namespaces.

These changes make kmux options consistent with the behavior of kubectl
when a context without a namespace is used.